Added requirements.txt, updated readme.md and added gitignore
